LLVM Analyze corrections
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@10948 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ SeriesView.h
@@ -43,8 +43,8 @@
 
 - (id)initWithFrame:(NSRect)frame seriesRows:(int)rows  seriesColumns:(int)columns;
 
-- (long)tag;
-- (void)setTag:(long)theTag;
+- (NSInteger)tag;
+- (void)setTag:(NSInteger)theTag;
 - (NSMutableArray *)imageViews;
 - (DCMView *)firstView;
 - (void)setImageViewMatrixForRows:(int)rows  columns:(int)columns;
